.experiment/finetuning/requirements.txt
torch
torchvision
tqdm
pillow
numpy
matplotlib
timm
scikit-learn